requirements.txt
flask
werkzeug
sqlalchemy
chromadb
google-generativeai
pypdf
python-dotenv
langchain
langchain-community
langchain-core
tiktoken
flask-cors
gunicorn
unstructured
pydantic
PyPDF2
google-cloud-speech
google-cloud-texttospeech
pyaudio
numpy
soundfile
scipy